Fixed a conversion on 32 bit platforms.
git-svn-id: http://yate.null.ro/svn/yate/trunk@5470 acf43c95-373e-0410-b603-e72c3f656dc1

+++ libs/yscript/jsobjects.cpp
@@ -1129,7 +1129,7 @@ bool JsDate::runNative(ObjList& stack, const ExpOperation& oper, GenObject* cont
     }
     else if (oper.name() == YSTRING("getTime")) {
 	// Returns the time in milliseconds since UNIX Epoch
-	ExpEvaluator::pushOne(stack,new ExpOperation(1000 * ((long int)m_time) + m_msec));
+	ExpEvaluator::pushOne(stack,new ExpOperation(1000 * ((long int)m_time) + (long int)m_msec));
     }
     else if (oper.name() == YSTRING("getTimezoneOffset")) {
 	// Returns the UTC to local difference in minutes, positive goes west
